requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.2
bleach==6.0.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer @ file:///private/tmp/python-charset-normalizer-20231123-4937-xhhqqy/charset-normalizer-3.3.2
click==8.1.8
comm==0.1.2
contourpy==1.3.1
cycler==0.12.1
datasets==3.3.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
executing==1.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastjsonschema==2.16.2
ffmpy==0.5.0
filelock==3.17.0
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gradio==5.16.0
gradio_client==1.7.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.4
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.5.1
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kagglehub==0.3.8
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.10.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.4.2
notebook==6.5.2
notebook_shim==0.2.2
numpy==2.2.3
orjson==3.10.15
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==3.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
propcache==0.2.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pycparser==2.21
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments @ file:///private/tmp/pygments-20231121-5079-e7uuzt/pygments-2.17.2
pyparsing==3.2.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.6
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
ruff==0.9.6
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
semantic-version==2.10.0
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
stack-data==0.6.2
starlette==0.45.3
sympy==1.13.1
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
tornado==6.2
tqdm==4.67.1
traitlets==5.9.0
transformers==4.48.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.2.0
urllib3 @ file:///private/tmp/python-urllib3-20231123-5269-2cyx6e/urllib3-2.1.0
uvicorn==0.34.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==14.2
widgetsnbextension==4.0.5
xxhash==3.5.0
yarl==1.18.3
